+++ tests/testthat/test-check_itemscale.R
@@ -1,4 +1,4 @@
-test_that("check_convergence", {
+test_that("check_itemscale", {
   skip_if_not_installed("parameters")
 
   set.seed(123)
@@ -24,5 +24,32 @@ test_that("check_convergence", {
     item_difficulty(d[out[[2]]$Item])$Difficulty,
     tolerance = 1e-4,
     ignore_attr = TRUE
+  )
+  expect_snapshot(print(out), variant = "windows")
+  comp <- parameters::closest_component(pca)
+  out2 <- check_itemscale(d, comp)
+  expect_equal(
+    out[[1]]$Mean,
+    out2[[1]]$Mean,
+    tolerance = 1e-4,
+    ignore_attr = TRUE
+  )
+  expect_equal(
+    out[[1]]$Difficulty,
+    out2[[1]]$Difficulty,
+    tolerance = 1e-4,
+    ignore_attr = TRUE
+  )
+  expect_error(
+    check_itemscale(d),
+    regex = "If `x` is a data"
+  )
+  expect_error(
+    check_itemscale(d, factor_index = 1:8),
+    regex = "`factor_index` must be of same"
+  )
+  expect_error(
+    check_itemscale(d, factor_index = factor(comp)),
+    regex = "`factor_index` must be numeric."
   )
 })